requirements.txt
Django==4.1
djangorestframework==3.13.1
Markdown==3.4.1
django-filter==22.1
django-rest-swagger==2.2.0
PyYAML==6.0
python-dotenv==0.20.0
django-cors-headers==3.13.0
gunicorn==20.1.0
Pillow==9.2.0
django-storages==1.13.1
boto3==1.24.59
